Isolated GraphQL schema delegates from implementation delegates
