+++ src/main/kotlin/com/prestongarno/ktq/hooks/Configurable.kt
@@ -1,6 +1,7 @@
 package com.prestongarno.ktq.hooks
 
 import com.prestongarno.ktq.ArgBuilder
+import com.prestongarno.ktq.DelegateProvider
 import com.prestongarno.ktq.QModel
 import com.prestongarno.ktq.QType
 import com.prestongarno.ktq.SchemaStub
