+++ src/main/kotlin/com/prestongarno/ktq/stubs/CustomScalars.kt
@@ -4,7 +4,7 @@ import com.prestongarno.ktq.CustomScalar
 import com.prestongarno.ktq.CustomScalarArgBuilder
 import com.prestongarno.ktq.SchemaStub
 import com.prestongarno.ktq.adapters.custom.QScalarMapper
-import com.prestongarno.ktq.hooks.DelegateProvider
+import com.prestongarno.ktq.DelegateProvider
 
 interface CustomScalarInitStub<T: CustomScalar> : SchemaStub {
   fun <U: QScalarMapper<A>, A> init(init: U): CustomStub<U, A>
